add scroll page mode
Currently, this can only be enabled by using expert configuration. Add following section to expert option to enable this feature.

```
[appearance]
read_style=scroll
```

This feature is only provided for testing. It is currently under development and not suitable for using. Once it is stable enough for beta test, an option will be added so it can be switched to without changing expert option.